Add bundle support to "burn remotepayload" command.
